Syntax extended to allow limited "large expressions."
A large expression is an expression that can contain statements. The
motivating case is a time block, which is now an expression that returns
the timing. But we can also make `if-elif-else` and `lambda` large
expressions. (The latter requires that `return` be allowed in `lambda`
now, hence the change to private/syntax.rkt.)

The contexts where large expressions are allowed are currently quite
limited: They are allowed on the right-hand sides of `let`s and
assignments, and after `return`s.